Resolved the issues.
